fileservice: add per file read lock to reduce duplicated IO (#11119)
fileservice: code clean-ups

fileservice: add IOLocks

fileservice: add io locks to LocalFS, S3FS

Approved by: @nnsgmsone
